add: host增加ip辅助字段，支持重复ip
