+++ conf/application-akka.conf
@@ -19,7 +19,7 @@ akka {
 
   remote {
     netty.tcp {
-      hostname = "172.19.3.114"
+      hostname = "172.19.3.106"
       port = 0
       send-buffer-size = 4m
       receive-buffer-size = 4m
