+++ conf/application-common.conf
@@ -155,4 +155,4 @@ file.exceeded.max_size=10485760
 file.extension.octet_list="key,jks,jar"
 
 # 项目版本
-app.version="6.1.0"
+app.version="6.1.1"
