+++ conf/application.conf
@@ -19,4 +19,4 @@ spirit.init = true
 
 # bugatti.system.baseUrl="http://bugatti2.dev.ofpay.com"
 
-bugatti.system.baseUrl="http://172.19.3.114:9000"
+bugatti.system.baseUrl="http://172.19.3.106:9000"
