+++ public/partials/admin/relation/relation-index.html
@@ -54,7 +54,7 @@
                 </div>
             </td>
             <td><span tooltip="{{rel.name}}">{{rel.name}}</span></td>
-            <td><span tooltip="{{rel.ip}}">{{rel.ip}}</span></td>
+            <td><span tooltip="{{rel.ip}} - {{rel.ipClash}}">{{rel.ip}}</span></td>
             <td><span tooltip="{{rel.hostIp}}">{{rel.hostIp}}</span></td>
             <td>
                 <a ng-click="unbind(rel.id, $index)" href tooltip="解绑[{{rel.id}}]"><i class="fa fa-chain-broken"></i></a>
